+++ include/ignition/math/MassMatrix3.hh
@@ -1040,6 +1040,24 @@ namespace ignition
         return this->SetMoi(R * L * R.Transposed());
       }
 
+      /// \brief Set inertial properties based on a material and
+      /// equivalent sphere.
+      /// \param[in] _mat Material that specifies a density. Uniform density
+      /// is used.
+      /// \param[in] _radius Radius of equivalent, uniform sphere.
+      /// \return True if inertial properties were set successfully.
+      public: bool SetFromSphere(const Material &_mat, const T _radius)
+      {
+        // Check that the density and _radius are strictly positive
+        if (_mat.Density() <= 0 || _radius <= 0)
+        {
+          return false;
+        }
+
+        double volume = (4.0/3.0) * IGN_PI * std::pow(_radius, 3);
+        return this->SetFromSphere(_mat.Density() * volume, _radius);
+      }
+
       /// \brief Set inertial properties based on mass and equivalent sphere.
       /// \param[in] _mass Mass to set.
       /// \param[in] _radius Radius of equivalent, uniform sphere.